feat: Add class-transformer and class-validator dependencies
This commit adds the "class-transformer" and "class-validator" dependencies to the project's package.json file. These libraries will be used for data transformation and validation in the codebase.

Additionally, the .env file in the projects/api directory has been updated to include a new PORT variable with a value of 52100.

In the musculoskeletal.entity.ts file, the column types for "name" and "type" have been changed from "string" to "text".

Lastly, an assertion check for REDIS_URL in the core.module.ts file has been commented out temporarily.